requirements.txt
alembic==1.18.4
bidict==0.23.1
black==26.3.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.28.0
flake8==7.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.62.1
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
idna==3.12
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
polars==1.40.0
polars-runtime-32==1.40.0
psutil==7.2.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pytokens==0.4.1
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.49
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.41
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.8
wsproto==1.3.2
WTForms==3.2.1
